PyRoomPlugin/requirements.txt
paho-mqtt>=1.3.1
PyQt5>=5.12
PyQt5-sip>=4.19
Unidecode>=1.1
